requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.4
pydantic-settings==2.3.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
email-validator==2.1.1
chromadb==0.4.24
sentence-transformers==2.7.0
numpy<2.0.0
cryptography==42.0.7
streamlit==1.36.0
requests==2.32.3
langchain==0.2.5
langchain-core==0.2.16
langchain-openai==0.1.8